+++ src/components/elements/buttons/AddCardButton.tsx
@@ -2,7 +2,7 @@ import { UserState, UserStateType } from "@/global-states/atoms";
 import Image from "next/image";
 import { useState } from "react";
 import { useRecoilValue } from "recoil";
-import { CreateRecruitModal } from "../recruitElement/CreateRecruitModal";
+import { CreateRecruitModal } from "../modal/CreateRecruitModal";
 
 export const AddCardButton = () => {
   const userAtomVal = useRecoilValue<UserStateType>(UserState);
@@ -11,29 +11,26 @@ export const AddCardButton = () => {
 
   const closeModal = () => {
     setIsOpen(false);
-  }
+  };
 
   const openModal = () => {
     setIsOpen(true);
-  }
+  };
   return (
     <>
-    <button
-      type="button"
-      onClick={openModal}
-    >
-      <div className="fixed bottom-5 right-14">
-      <div className="">
-        <Image src="/plus.gif" alt="" width={150} height={150} />
-      </div>
-    </div>
-    </button>
+      <button type="button" onClick={openModal}>
+        <div className="fixed bottom-5 right-14">
+          <div className="">
+            <Image src="/plus.gif" alt="" width={150} height={150} />
+          </div>
+        </div>
+      </button>
 
-    <CreateRecruitModal
-      isOpen={isOpen}
-      closeModal={closeModal}
-      user={userAtomVal?.uid}
-    />
+      <CreateRecruitModal
+        isOpen={isOpen}
+        closeModal={closeModal}
+        user={userAtomVal?.uid}
+      />
     </>
   );
 };
