+++ src/components/elements/buttons/NewOrder.tsx
@@ -1,0 +1,7 @@
+export const NewOrder = () => {
+  return (
+    <button className="bg-white hover:bg-tertiary-color font-zen py-2 px-4 mr-2 text-lg rounded-2xl inline-flex items-center">
+      新着順
+    </button>
+  );
+};
